nuvion_app/inference/requirements-zsad.txt
torch
transformers<5
Pillow
